app/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pdfplumber==0.11.5
python-docx==1.1.2
openpyxl==3.1.5
